Create README.md
This project contains two datasets. Train dataset should be used to train the data and Test dataset should be used to test the trained model. 
The data consists of observations on loan status (paidoff or collection) along with individual's features. 
My goal is to identify what is the best classification algorithm to predict an individual's loan status.